Move logic for download handling into the `NotebookOutputWebview` class intead of the notebook webview service.
